# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Cryptocoin Exodo",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "https://criptocoin.rj.r.appspot.com",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Efetua o login verificando se os campos estão validos",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "any"
                },
                "password": {
                  "type": "string",
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Solicitação Inválida"
          }
        }
      }
    },
    "/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Efetua o registro, retorna o qrcode,secret32 e o usuario cadastrado",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "any"
                },
                "email": {
                  "type": "string",
                  "example": "any"
                },
                "password": {
                  "type": "string",
                  "example": "any"
                },
                "cpf": {
                  "type": "string",
                  "example": "any"
                },
                "rg": {
                  "type": "string",
                  "example": "any"
                },
                "referenciadoPor": {
                  "type": "string",
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "422": {}
        }
      }
    },
    "/verify": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Efetua a verificação do token 2FA do cadastro",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token2FA": {
                  "type": "string",
                  "example": "any"
                },
                "email": {
                  "type": "string",
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/validate": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Efetua a validação do token 2FA do login",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token2FA": {
                  "type": "string",
                  "example": "any"
                },
                "email": {
                  "type": "string",
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/coin": {
      "post": {
        "tags": [
          "Moeda"
        ],
        "description": "Cadastro de uma moeda e retorna ela",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "any"
                },
                "sigla": {
                  "type": "string",
                  "example": "any"
                },
                "valor": {
                  "type": "string",
                  "example": "any"
                },
                "quantidade": {
                  "type": "string",
                  "example": "any"
                },
                "rendimento": {
                  "type": "string",
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Não autorizado"
          },
          "422": {}
        }
      },
      "get": {
        "tags": [
          "Moeda"
        ],
        "description": "Retorna todas as moedas",
        "parameters": [],
        "responses": {
          "401": {
            "description": "Não autorizado"
          },
          "422": {}
        }
      }
    },
    "/coin/{nome}": {
      "get": {
        "tags": [
          "Moeda"
        ],
        "description": "Retorna uma moeda",
        "parameters": [
          {
            "name": "nome",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Não autorizado"
          },
          "422": {}
        }
      }
    },
    "/coin/{sigla}": {
      "delete": {
        "tags": [
          "Moeda"
        ],
        "description": "Deleta uma moeda",
        "parameters": [
          {
            "name": "sigla",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Não autorizado"
          },
          "422": {}
        }
      }
    },
...
  },
...
}